move token provider to private header file
