+++ source/token_provider_sso_profile.c
@@ -7,6 +7,7 @@
 
 #include <aws/auth/private/aws_profile.h>
 #include <aws/auth/private/credentials_utils.h>
+#include <aws/auth/private/sso_token_providers.h>
 #include <aws/auth/private/sso_token_utils.h>
 #include <aws/common/clock.h>
